+++ trello/admin.py
@@ -6,5 +6,5 @@ admin.site.register(List)
 admin.site.register(Board)
 admin.site.register(BoardMembers)
 admin.site.register(Card)
-admin.site.register(BoardInvite)
 admin.site.register(CardImage)
+admin.site.register(BoardInvite)
