--- trello/static/css/ultraultra.css
@@ -1,253 +1,0 @@
-body{
-    font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Oxygen, Ubuntu, Cantarell, 'Open Sans', 'Helvetica Neue', sans-serif;
-    background-image: url("../images/charlie-foster-yPaskTQiNjA-unsplash.jpg");
-    opacity: 5; 
-}
-
-.board-header{
-    height: 70px;
-}
-
-.board-row{
-    height: 70px;
-}
-
-#add-board{
-    color: #ffffff;
-    font-size: 25px;
-    padding: 0px;
-    height: 10px;
-}
-
-#create-board{
-    padding-top: 0px;
-    height: 10px;
-}
-
-.input-card{
-    width: 245px;
-    height: 30px;
-    border: none;
-    padding-left: 8px;
-    font-size: 14px;
-}
-
-.input-member-email{
-    width: 245px;
-    padding-left: 8px;
-    margin-bottom: -5px;
-    font-size: 14px;
-}
-
-#trello-header{
-    margin-top: -43px;
-}
-
-.create-card{
-    margin-top: -5px;
-}
-
-.page-header{
-    height: 55px;
-}
-
-.page-header a{
-    color: #ffffff;
-    float: right;
-}
-
-
-#board_dropdown{
-    margin-top: -44px;
-    margin-left: 75%;
-}
-
-.modal{
-    width: 50%;
-}
-
-/*.card{
-    width: 220px;
-    font-size: 10px;
-}*/
-
-#logout{
-    margin-top: -45px;
-    padding-bottom: 0;
-    width: 90px;
-    float: right;
-    height: 40px;
-}
-
-.card {
-    background-color: #DCDCDC;
-    border: none;
-    width: 272px;
-    min-width: 260px;
-    align-self: flex-start; 
-     
-}
-.card-title {
-    height: 40px;
-    padding-left: 15px;
-    padding-right: 5px;
-    color: #172b4d;
-}
-
-.card-body {
-    margin-top: -17px;
-    margin-left: 7px;
-    display: inline-block;
-    padding: 3px;
-    font-size: 16px;
-    width: 250px;
-}
-
-
-.closeCard {
-    background: none; 
-}
-
-#list-dropdown{
-    margin-top: -15px;
-    color: #172b4d;
-    font-weight: bold;
-    height: 10px;
-}
-
-#dropdownMenuButton{
-    font-size: 20px;
-}
-
-.ltitle{
-    height: 30px;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: clip;
-}
-
-#list{
-    color: #172b4d;
-    font-size: 14px;
-    font-weight: 500;
-    line-height: 20px;
-}
-
-.addcard{
-    color: #172b4d;
-    font-weight: 400;
-    font-size: 14px;
-    background: #fafafa;
-    height: 32px;
-    padding: 9px;
-}
-
-.container-fluid{
-    display: inline-block;
-}
-
-#id_card_title{
-    color: #172b4d;
-    font-weight: 400;
-    font-size: 14px;
-    border: none;
-    height: 40px;
-    border-radius: 0;
-}
-
-#list-container{
-    height: 35px;
-    width: 250px;
-}
-
-.glyphicon-pencil{
-    font-size: 10px;
-    color: #484848;
-}
-
-#description{
-    padding: 1px;
-}
-
-.glyphicon-trash{
-    font-size: 16px;
-    padding-right: 0;
-}
-
-.glyphicon-folder-open{
-    width: 30px;
-    font-size: 20px;
-}
-
-#board-header{
-}
-
-#board_title{
-    height: 20px;
-}
-
-#card-content{
-    width: 570px;
-    height: 680px;
-}
-
-#modal-card{
-    margin: auto;
-}
-
-.header-board{
-    margin-top: 14px;
-    margin-left: 120px;
-}
-
-#archive{
-    margin-top: -1px;
-    margin-left: 8px;
-}
-
-#show-details{
-    background-color: rgba(9, 30, 66, .05)
-}
-
-#archive-card{
-    background-color: rgba(9, 30, 66, .05)
-}
-
-.col #dropdownMenuButton{
-    font-size: 15px;
-    background-color: rgba(9, 30, 66, .05);
-}
-
-#updateBoardForm{
-    display: none;
-}
-
-#updateListForm{
-    display: none;
-}
-
-#input-list{
-    display: none;
-}
-
-#updateCardForm{
-    display: none;
-}
-
-#inviteDropdownMenu{
-    text-align: center;
-    width: 300px;
-}
-
-#card_description{
-    display: none;
-}
-
-.update-error{
-    display: none; 
-    margin: auto; 
-    font-weight: normal; 
-    color:#FF0000; 
-    font-size: 12px;
-}
-
