+++ trello/templates/trello/base.html
@@ -5,7 +5,7 @@
 
         <link rel="stylesheet" href="{% static 'node_modules/bootstrap/dist/css/bootstrap.css' %}">
         <link rel="stylesheet" href="{% static 'node_modules/bootstrap3/dist/css/bootstrap.min.css' %}">
-        <link rel="stylesheet" href="{% static 'css/ultraultra.css' %}">
+        <link rel="stylesheet" href="{% static 'css/ultraultrage.css' %}">
     </head>
     <body>
         {% include 'trello/modals.html' %}
@@ -17,7 +17,7 @@
 
         <script src="{% static 'node_modules/jquery/dist/jquery.js' %}"></script>
         <script src="{% static 'node_modules/jquery-ui-1.12.1.custom/jquery-ui.js' %}"></script>
-        <script src="{% static 'js/ultraultrag.js' %}"></script> 
+        <script src="{% static 'js/ultraultrage.js' %}"></script> 
         <script src="{% static 'node_modules/popper.js/dist/umd/popper.min.js' %}"></script>
         <script src="{% static 'node_modules/bootstrap/dist/js/bootstrap.js' %}"></script>            
     </body>
