+++ trello/templates/trello/list.html
@@ -3,13 +3,23 @@
 {% block content %}
     <div class="container-fluid mt-3">
         <h2 class="mb-5">List of Boards</h2>
+        <h5 class="mt-3">Personal Boards</h5>
         {% for b in board %}
             <ul class="list-unstyled ml-3"><li><a class="text-dark" href="{% url 'board' id=b.id %}">{{ b.title }} </a></li></ul>
         {% endfor %}
+        <h5 class="mt-5">Invite Boards</h5>
         {% for boards in board_member %}
             {% if boards.deactivate == False %}
-                <ul class="list-unstyled ml-3"><li><a class="text-dark" href="{% url 'board' boards.board.id %}">{{ boards.board.title }}</a></li></ul>
+                {% if boards.owner == False %}
+                    <ul class="list-unstyled ml-3"><li><a class="text-dark" href="{% url 'board' boards.board.id %}">{{ boards.board.title }}</a></li></ul>
+                {% endif %}
             {% endif %}
         {% endfor %}
+        <div class="container">
+            KOKOKOKOKO
+            {% for i in card_image %}
+                {{ i.image }} <img src="{{ i.image.url }}" class="img-responsive">
+            {% endfor %}
+        </div>
     </div>
 {% endblock %}
