+++ trello/templates/trello/modals.html
@@ -38,7 +38,7 @@
                     <span class="p-0" aria-hidden="true">&times;</span>
                 </button>
             </div>
-            <div class="modal-body pt-0 mt-0">
+            <div class="modal-body pt-0 mt-4">
                 <p>Loading...</p>
             </div>
         </div>
